% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{clr_transform}
\alias{clr_transform}
\title{Apply centered log-ratio (CLR) transformation to compositional data}
\usage{
clr_transform(X, pseudocount = 1)
}
\arguments{
\item{X}{Matrix where rows are clusters and columns are samples}

\item{pseudocount}{Small value added to avoid log(0), defaults to 1}
}
\value{
Matrix of CLR-transformed values
}
\description{
Apply centered log-ratio (CLR) transformation to compositional data
}
\details{
The CLR transformation is computed as:
CLR(x) = log(x) - mean(log(x))
where x is a composition vector
}
