% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_genes_and_Peaks.R
\name{get_binned}
\alias{get_binned}
\title{Process peak matrix data with options for RNA and non-RNA normalization}
\usage{
get_binned(mat_atac_aggregated, pseudotime_atac_agg)
}
\arguments{
\item{mat_}{Input matrix to process}

\item{RNA}{Logical indicating if data is RNA (default: TRUE)}
}
\value{
Processed matrix with:
\itemize{
\item Cleaned rownames (prefixes removed)
\item Optional CPM normalization for non-RNA data
}
}
\description{
Process peak matrix data with options for RNA and non-RNA normalization
}
