% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_genes_and_Peaks.R
\name{get_normed_binned_each_cellType}
\alias{get_normed_binned_each_cellType}
\title{Process and bin data for each cell type}
\usage{
get_normed_binned_each_cellType(ptime, cellTypes, pseudotime, RNA = T)
}
\arguments{
\item{ptime}{List containing pseudobulk data in ptime$pb.ls$all}

\item{cellTypes}{Vector of cell type names to process}

\item{pseudotime}{Named vector of pseudotime values for each sample}

\item{RNA}{Logical indicating if input is RNA data (default: TRUE)}
}
\value{
List of binned data for each cell type after normalization
}
\description{
Process and bin data for each cell type
}
\details{
\enumerate{
\item Finds common samples across all cell types
\item For each cell type:
\itemize{
\item Normalizes data using get_processed_peaks()
\item Bins normalized data using get_binned()
}
}
}
