% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_genes_and_Peaks.R
\name{get_processed_peaks}
\alias{get_processed_peaks}
\title{Process peak matrix data with options for RNA and non-RNA normalization}
\usage{
get_processed_peaks(mat_, RNA = T)
}
\arguments{
\item{mat_}{Input matrix to process}

\item{RNA}{Logical indicating if data is RNA. If FALSE, performs CPM normalization (default: TRUE)}
}
\value{
Processed matrix with cleaned rownames and optional CPM normalization
}
\description{
Processes a matrix of peak data by cleaning rownames and optionally performing
counts per million (CPM) normalization for non-RNA data
}
\details{
For RNA data, only cleans rownames by removing prefixes
For non-RNA data, additionally performs CPM normalization:
\enumerate{
\item Calculates column sums
\item Divides each column by its sum
\item Multiplies by 10^6 for CPM
}
}
