% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_sample_pb}
\alias{get_sample_pb}
\title{Generates pseudobulk profiles from single-cell data with preprocessing options
Parameters:}
\usage{
get_sample_pb(
  s,
  pt,
  HVG = T,
  qnorm = F,
  combat = F,
  batch = NULL,
  preserve_zero = T
)
}
\arguments{
\item{s}{Single-cell expression matrix (genes x cells)}

\item{pt}{Vector of sample/patient IDs for each cell}

\item{HVG}{Whether to select highly variable genes (default: TRUE)}

\item{qnorm}{Whether to perform quantile normalization (default: FALSE)}

\item{combat}{Whether to apply ComBat batch correction (default: FALSE)}

\item{batch}{Batch labels for ComBat correction}

\item{preserve_zero}{Whether to preserve zeros in quantile normalization (default: TRUE)
Returns:
List containing:
\itemize{
\item all: Full preprocessed pseudobulk matrix
\item hvg: Subset of highly variable genes (if HVG=TRUE)
}}
}
\description{
Generates pseudobulk profiles from single-cell data with preprocessing options
Parameters:
}
