% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{getSeuratProcessed}
\alias{getSeuratProcessed}
\title{Create and process a Seurat object from log2-transformed expression matrix}
\usage{
getSeuratProcessed(mat_log2, nfeature_)
}
\arguments{
\item{mat_log2}{Log2-transformed expression matrix}

\item{nfeature_}{Number of variable features to select}
}
\value{
Processed Seurat object containing:
\itemize{
\item Normalized data
\item Original log2 data in RNA assay
\item Selected variable features
}
}
\description{
Create and process a Seurat object from log2-transformed expression matrix
}
