% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{getZscore}
\alias{getZscore}
\title{Calculate standardized variance z-scores for highly variable features}
\usage{
getZscore(obj_)
}
\arguments{
\item{obj_}{Seurat object containing gene expression data}
}
\value{
Named vector of standardized variance z-scores, with gene names as names
}
\description{
Calculate standardized variance z-scores for highly variable features
}
