% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genes_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Create pseudotime-ordered gene expression heatmaps}
\usage{
plot_heatmap(list_norm, genes_, clu_genes_, pseudotime_)
}
\arguments{
\item{list_norm}{List of normalized expression matrices for different cell types}

\item{genes_}{Vector of gene names to plot}

\item{clu_genes_}{Named vector of cluster assignments for genes}

\item{pseudotime_}{Named vector of pseudotime values for samples}
}
\description{
Create pseudotime-ordered gene expression heatmaps
}
\details{
The function:
\itemize{
\item Orders samples by pseudotime
\item Orders genes by cluster
\item Scales expression values to z-scores
\item Caps z-scores at \link{-1,1}
\item Creates separate heatmaps for each cell type
\item Annotates genes by cluster assignment
}
}
