% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{remove_batch_effect}
\alias{remove_batch_effect}
\title{Remove batch effects from expression data using linear modeling}
\usage{
remove_batch_effect(dat, batch, adj_cov = NULL)
}
\arguments{
\item{dat}{Matrix of expression data (features x samples)}

\item{batch}{Factor indicating batch assignments for each sample}

\item{adj_cov}{Additional covariates matrix to preserve during batch correction}
}
\value{
Batch-corrected expression matrix with same dimensions as input
}
\description{
Remove batch effects from expression data using linear modeling
}
\details{
Implements batch correction by:
\enumerate{
\item Fitting linear model with batch effects
\item Estimating batch parameters
\item Subtracting batch effects while preserving biological variation
}
}
